Minor: Consolidate UDF tests (#7704)
* Minor: Consolidate user defined functions

* cleanup

* move more tests

* more

* cleanup use
